Fast-and-dirty Apache "AllowEncodedSlashes No" support
Some web hosting services have Apache "AllowEncodedSlashes" settings
set to "Off". They rejects %2F ("/") on unix-like systems by throwing
a fake 404 (Not found) error. The same for %5C on Windows systems.

Exemple of hosting services includes Redhat Openshift cloud.

This commit is a fast-and-dirty fix for my unix-like case.
It forces slash characters to be sent as is for HTTP requests.